account activation

+++ app/Http/Controllers/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Auth;
+use Mail;
 
 class UsersController extends Controller
 {  
@@ -14,7 +15,7 @@ class UsersController extends Controller
      */
     public function __construct(){
         $this->middleware('auth',[
-            'except'=>['show','create','store','index']
+            'except'=>['show','create','store','index','confirmEmail']
             ]);
 
         $this->middleware('guest', [
@@ -79,10 +80,10 @@ class UsersController extends Controller
             'password' => bcrypt($request->password),
             ]);
 
-        //消息提醒
-        Auth::login($user);
-        session()->flash('success', '欢迎，您将在这里开启一段新的旅程~');
-    	return redirect()->route('users.show',[$user->id]);
+        //发送激活邮件
+        $this -> sendEmailConfirmationTo($user);
+        session()->flash('success', '验证邮件已经发送到你的注册邮箱上，请注意查收。');
+    	return redirect('/');
 
     }
 
@@ -148,7 +149,7 @@ class UsersController extends Controller
      */
      public function destroy(User $user){
 
-        
+        $this->authorize('destroy',$user);
 
     	$user->delete();
 
@@ -156,6 +157,53 @@ class UsersController extends Controller
 
         return back();
 
+    }
+
+    /**
+     * [sendEmailConfirmationTo description]
+     * 
+     * @param  [type] $user [description]
+     * @return [type]       [description]
+     */
+    protected function sendEmailConfirmationTo($user){
+
+        $view = 'emails.confirm';
+        $data = compact('user');
+        $from = 'root@email.com';
+        $name = 'root';
+        $to = $user->email;
+        $subject = "感谢注册sample应用！ 请确认你的邮箱。";
+
+        Mail::send($view,$data,function($message) use($from,$name,$to,$subject){
+
+            $message->from($from,$name)->to($to)->subject($subject);
+        });
+    }
+
+    /**
+     * [confirmEmail description]
+     * 
+     * @param  [type] $token [description]
+     * @return [type]        [description]
+     */
+    public function confirmEmail($token){
+
+        $user = User::where('activation_token',$token)->firstOrFail();
+
+        if($token != $user->activation_token){
+
+            session()->flash('warning','对不起，激活失败......');
+
+            return redirect('/');
+
+        }
+        $user->activated = true;
+        $user->activation_token = '';
+        $user->save();
+
+        Auth::login($user);
+        session()->flash('success','恭喜你，激活成功！');
+        return redirect()->route('users.show',[$user]);
     }
     
 }
